requirements.txt
faiss_cpu==1.11.0.post1
fastapi==0.116.1
langchain==0.3.27
langgraph==0.6.7
numpy==2.3.2
openai==1.107.0
Pillow==11.3.0
pydantic==2.11.7
python-dotenv==1.1.1
rapidfuzz==3.14.1
streamlit==1.45.1
torch==2.7.1
transformers==4.54.1
twilio==8.4.1
